% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois.nbinom.hamim.R
\name{distribusi.hamim}
\alias{distribusi.hamim}
\title{Goodness-of-fit for Poisson / NegBin with two-sided pooling (EN/ID)}
\usage{
distribusi.hamim(
  data,
  distrib = c("poisson", "negbinom"),
  method = c("MLE", "k.langsung", "manual.iterasi", "log.iterasi"),
  ex_min = 5,
  min_expected = NULL,
  plot = TRUE,
  tail = FALSE,
  tail_from = NULL,
  pool = c("auto", "none")
)
}
\arguments{
\item{data}{data.frame dengan kolom X (cacah, integer) dan Fx (frekuensi).}

\item{distrib}{c("poisson","negbinom")}

\item{method}{c("MLE","k.langsung","manual.iterasi","log.iterasi")}

\item{ex_min}{numeric; ambang Ex}

\item{min_expected}{alias lama untuk ex_min}

\item{plot, tail, tail_from, pool}{lihat help}
}
\value{
list: estimasi & tabel hasil
}
\description{
Goodness-of-fit for Poisson / NegBin with two-sided pooling (EN/ID)
}
\examples{
dat <- data.frame(X = 0:6, Fx = c(114, 25, 15, 10, 6, 5, 2))
prev <- distribusi.hamim(dat, distrib = "poisson", pool = "none", tail = TRUE, plot = FALSE)
sum(prev$tabel$Px)
gof  <- distribusi.hamim(dat, distrib = "poisson", ex_min = 3, pool = "auto", tail = TRUE, plot = FALSE)
gof$p_val; gof$df
nb   <- distribusi.hamim(dat, distrib = "negbinom", method = "log.iterasi",
                         ex_min = 5, pool = "auto", tail = TRUE, plot = FALSE)
nb$k; nb$mu; nb$p_val
}
