% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois.nbinom.hamim.R
\name{info.distribusi.hamim}
\alias{info.distribusi.hamim}
\alias{info.pois.nbinom.hamim}
\title{Info singkat indeks dispersi & GOF (EN/ID)}
\usage{
info.distribusi.hamim()

info.pois.nbinom.hamim()
}
\description{
Info singkat indeks dispersi & GOF (EN/ID)
}
